refactor: move probes server in it's own package
