+++ pkg/probes/true.go
@@ -1,4 +1,4 @@
-package handlers
+package probes
 
 func True() bool {
 	return true
